+++ lib/model/map/ArcO2valMapBuilder.php
@@ -63,7 +63,7 @@ class ArcO2valMapBuilder {
 
 		$tMap->addPrimaryKey('ID', 'Id', 'int', CreoleTypes::INTEGER, true, null);
 
-		$tMap->addColumn('CID', 'Cid', 'int', CreoleTypes::SMALLINT, true, null);
+		$tMap->addColumn('CID', 'Cid', 'int', CreoleTypes::INTEGER, true, null);
 
 		$tMap->addColumn('MISC', 'Misc', 'boolean', CreoleTypes::BOOLEAN, true, null);
